move contents of __init__.py to commands.py, connection.py
